+++ src/agent.rs
@@ -13,26 +13,39 @@ pub enum Token {
     /// Call the [`Keyboard::key`] fn with the given key and direction
     #[cfg_attr(feature = "serde", serde(alias = "K"))]
     #[cfg_attr(feature = "serde", serde(alias = "k"))]
-    Key(Key, Direction),
+    Key(
+        Key,
+        #[cfg_attr(feature = "serde", serde(default))] Direction,
+    ),
     /// Call the [`Keyboard::raw`] fn with the given keycode and direction
     #[cfg_attr(feature = "serde", serde(alias = "R"))]
     #[cfg_attr(feature = "serde", serde(alias = "r"))]
-    Raw(u16, Direction),
+    Raw(
+        u16,
+        #[cfg_attr(feature = "serde", serde(default))] Direction,
+    ),
     /// Call the [`Mouse::button`] fn with the given mouse button and direction
     #[cfg_attr(feature = "serde", serde(alias = "B"))]
     #[cfg_attr(feature = "serde", serde(alias = "b"))]
-    Button(Button, Direction),
+    Button(
+        Button,
+        #[cfg_attr(feature = "serde", serde(default))] Direction,
+    ),
     /// Call the [`Mouse::move_mouse`] fn. The first i32 is the value to move on
     /// the x-axis and the second i32 is the value to move on the y-axis. The
     /// coordinate defines if the given coordinates are absolute of relative to
     /// the current position of the mouse.
     #[cfg_attr(feature = "serde", serde(alias = "M"))]
     #[cfg_attr(feature = "serde", serde(alias = "m"))]
-    MoveMouse(i32, i32, Coordinate),
+    MoveMouse(
+        i32,
+        i32,
+        #[cfg_attr(feature = "serde", serde(default))] Coordinate,
+    ),
     /// Call the [`Mouse::scroll`] fn.
     #[cfg_attr(feature = "serde", serde(alias = "S"))]
     #[cfg_attr(feature = "serde", serde(alias = "s"))]
-    Scroll(i32, Axis),
+    Scroll(i32, #[cfg_attr(feature = "serde", serde(default))] Axis),
 }
 
 pub trait Agent