+++ CHANGES.md
@@ -6,6 +6,7 @@ win, macOS: Allow marking events that were created by enigo. Have a look at the 
 
 ## Fixed
 win: Respect the language of the current window to determine the which scancodes to send
+win: Send the virtual key and its scan code in the events to work with programs that only look at one of the two
 
 # 0.2.0