+++ CHANGES.md
@@ -4,6 +4,7 @@
 ## Added
 all: Serialized tokens can be less verbose because serde aliases were added
 win, macOS: Allow marking events that were created by enigo. Have a look at the additional field of the `Settings` struct and the new method `get_marker_value` of the `enigo` struct (only available on Windows and macOS)
+all: The enums `Button`, `Direction`, `Axis` and `Coordinate` implement `Default`
 
 ## Fixed